Splitted utils.js into 3 files by thier specification
