+++ src/utils/task.js
@@ -1,0 +1,38 @@
+const getCurrentDate = () => {
+  const currentDate = new Date();
+
+  currentDate.setHours(23, 59, 59, 999);
+
+  return new Date(currentDate);
+};
+
+export const isTaskExpired = (dueDate) => {
+  if (dueDate === null) {
+    return false;
+  }
+
+  const currentDate = getCurrentDate();
+
+  return currentDate > dueDate.getTime();
+};
+
+export const isTaskExpiringToday = (dueDate) => {
+  if (dueDate === null) {
+    return false;
+  }
+
+  const currentDate = getCurrentDate();
+
+  return currentDate.getTime() === dueDate.getTime();
+};
+
+export const isTaskRepeating = (repeating) => {
+  return Object.values(repeating).some(Boolean);
+};
+
+export const humanizeTaskDueDate = (dueDate) => {
+  return dueDate.toLocaleString(`en-US`, {
+    day: `numeric`,
+    month: `long`
+  });
+};
